Refactor measurement models
