+++ gc_registry/measurement/schemas.py
@@ -1,5 +1,6 @@
 import datetime
 
+from pydantic import BaseModel
 from sqlmodel import Field
 
 from gc_registry import utils
@@ -16,3 +17,20 @@ class MeasurementReportBase(utils.ActiveRecord):
         description="Indicates whether the usage is gross or net of any losses in the system.",
     )
     is_deleted: bool = Field(default=False)
+
+
+class MeasurementReportRead(MeasurementReportBase):
+    id: int
+
+
+class MeasurementReportUpdate(BaseModel):
+    device_id: int | None = None
+    interval_start_datetime: datetime.datetime | None = None
+    interval_end_datetime: datetime.datetime | None = None
+
+
+class MeasurementSubmissionResponse(BaseModel):
+    message: str
+    first_reading_datetime: datetime.datetime
+    last_reading_datetime: datetime.datetime
+    total_device_usage: int
